+++ ext/gettextpo/gettextpo.c
@@ -487,6 +487,11 @@ UPDATE_FLAG (workflow);
 
 /**
  * Document-method: workflow_flag_iterator
+ * call-seq: workflow_flag_iterator -> GettextPO::FlagIterator
+ *
+ * You may find it handy to use the #workflow_flag_iterator method to
+ * iterate over the sticky flag.  This is for more versatile
+ * manipulation.
  */
 FLAG_ITER (workflow);
 
@@ -565,6 +570,11 @@ UPDATE_FLAG (sticky);
 
 /**
  * Document-method: sticky_flag_iterator
+ * call-seq: sticky_flag_iterator -> GettextPO::FlagIterator
+ *
+ * You may find it handy to use the #sticky_flag_iterator method to
+ * iterate over the sticky flag.  This is for more versatile
+ * manipulation.
  */
 FLAG_ITER (sticky);
 #endif